Flytter all kode fra ForvalterController til ForvaltningController. 1 forvalterendepunkt er nok
